Separating logic for container and component
